+++ ark-common/src/main/java/au/org/theark/core/model/lims/entity/BiodataFieldLov.java
@@ -20,7 +20,7 @@ import au.org.theark.core.model.Constants;
 public class BiodataFieldLov implements java.io.Serializable
 {
 
-	private int		id;
+	private Long		id;
 	private int		listId;
 	private String	value;
 	private int		order;
@@ -29,7 +29,7 @@ public class BiodataFieldLov implements java.io.Serializable
 	{
 	}
 
-	public BiodataFieldLov(int id, int listId, String value, int order)
+	public BiodataFieldLov(Long id, int listId, String value, int order)
 	{
 		this.id = id;
 		this.listId = listId;
@@ -38,14 +38,14 @@ public class BiodataFieldLov implements java.io.Serializable
 	}
 
 	@Id
-	@SequenceGenerator(name="biodatafieldlov_generator", sequenceName="BIODATAFIELDLOV_SEQUENCE")
-	@GeneratedValue(strategy=GenerationType.AUTO, generator = "biodatafieldlov_generator")
-	public int getId()
+	@SequenceGenerator(name = "biodatafieldlov_generator", sequenceName = "BIODATAFIELDLOV_SEQUENCE")
+	@GeneratedValue(strategy = GenerationType.AUTO, generator = "biodatafieldlov_generator")
+	public Long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
