+++ ark-common/src/main/java/au/org/theark/core/model/lims/entity/BiodataLovList.java
@@ -20,7 +20,7 @@ import au.org.theark.core.model.Constants;
 public class BiodataLovList implements java.io.Serializable
 {
 
-	private int		id;
+	private Long		id;
 	private String	name;
 	private String	description;
 
@@ -28,13 +28,13 @@ public class BiodataLovList implements java.io.Serializable
 	{
 	}
 
-	public BiodataLovList(int id, String name)
+	public BiodataLovList(Long id, String name)
 	{
 		this.id = id;
 		this.name = name;
 	}
 
-	public BiodataLovList(int id, String name, String description)
+	public BiodataLovList(Long id, String name, String description)
 	{
 		this.id = id;
 		this.name = name;
@@ -42,14 +42,14 @@ public class BiodataLovList implements java.io.Serializable
 	}
 
 	@Id
-	@SequenceGenerator(name="biodatalovlist_generator", sequenceName="BIODATALOVLIST_SEQUENCE")
-	@GeneratedValue(strategy=GenerationType.AUTO, generator = "biodatalovlist_generator")
-	public int getId()
+	@SequenceGenerator(name = "biodatalovlist_generator", sequenceName = "BIODATALOVLIST_SEQUENCE")
+	@GeneratedValue(strategy = GenerationType.AUTO, generator = "biodatalovlist_generator")
+	public Long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
