+++ ark-common/src/main/java/au/org/theark/core/model/lims/entity/Flag.java
@@ -21,7 +21,7 @@ import au.org.theark.core.model.Constants;
 public class Flag implements java.io.Serializable
 {
 
-	private int			id;
+	private Long		id;
 	private String		timestamp;
 	private Integer	deleted;
 	private String		domain;
@@ -32,7 +32,7 @@ public class Flag implements java.io.Serializable
 	{
 	}
 
-	public Flag(int id, String domain, int referenceId, String user)
+	public Flag(Long id, String domain, int referenceId, String user)
 	{
 		this.id = id;
 		this.domain = domain;
@@ -40,7 +40,7 @@ public class Flag implements java.io.Serializable
 		this.user = user;
 	}
 
-	public Flag(int id, Integer deleted, String domain, int referenceId, String user)
+	public Flag(Long id, Integer deleted, String domain, int referenceId, String user)
 	{
 		this.id = id;
 		this.deleted = deleted;
@@ -50,14 +50,14 @@ public class Flag implements java.io.Serializable
 	}
 
 	@Id
-	@SequenceGenerator(name="flag_generator", sequenceName="FLAG_SEQUENCE")
-	@GeneratedValue(strategy=GenerationType.AUTO, generator = "flag_generator")
-	public int getId()
+	@SequenceGenerator(name = "flag_generator", sequenceName = "FLAG_SEQUENCE")
+	@GeneratedValue(strategy = GenerationType.AUTO, generator = "flag_generator")
+	public Long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
